% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ind_Evaluation_30112018.R
\name{Delta_jki}
\alias{Delta_jki}
\title{Calculates the factor Delta_jki}
\usage{
Delta_jki(x, i, K)
}
\arguments{
\item{x}{a dataset.}

\item{i}{an observation index.}

\item{K}{the total number of linguistics in a sub-item.}
}
\value{
The response matrix of binary values (0 or 1) related to the answers of a particular dataset for its corresponding sub-items.
}
\description{
Calculates the factor Delta_jki
}
