% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FANOVA_Distance_06122018.R
\name{FANOVA.summary}
\alias{FANOVA.summary}
\title{Prints the summary of the estimation of a FANOVA metric-based model}
\usage{
FANOVA.summary(res)
}
\arguments{
\item{res}{a result of a call of the function FANOVA, where method = "distance".}
}
\value{
Returns a list of summary statistics of the estimated model given in res, shown in a FANOVA table. In addition, the F-statistics with their p-values, and the decision are given.
}
\description{
Prints the summary of the estimation of a FANOVA metric-based model
}
