% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMANOVA_Distance_10122018.R
\name{FMANOVA.interaction.summary}
\alias{FMANOVA.interaction.summary}
\title{Prints the summary of the estimation of the interaction in a Mult-FANOVA metric-based model}
\usage{
FMANOVA.interaction.summary(res)
}
\arguments{
\item{res}{a result of a call of the function FMANOVA, where method = "distance".}
}
\value{
Returns a list of summary statistics of the estimated model given in res, shown in a FANOVA table. In addition, the F-statistics with their p-values, and the decision are given.
}
\description{
Prints the summary of the estimation of the interaction in a Mult-FANOVA metric-based model
}
