% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy_Sample_Variance_Exact_31102018.R
\name{Fuzzy.exact.variance.poly.left}
\alias{Fuzzy.exact.variance.poly.left}
\title{Gives the polynomial forms of the numerical alpha-cuts modelling the exact variance}
\usage{
Fuzzy.exact.variance.poly.left(data.fuzzified, breakpoints = 100)
}
\arguments{
\item{data.fuzzified}{a fuzzification matrix constructed by a call to the function FUZZ or the function GFUZZ, 
or a similar matrix. No NA are allowed.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A table composed by the coefficients of the second order equations of the left side, given at the corresponding definitions domains.
}
\description{
Gives the polynomial forms of the numerical alpha-cuts modelling the exact variance
}
