% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy_Sample_Variance_Exact_31102018.R
\name{Fuzzy.exact.variance}
\alias{Fuzzy.exact.variance}
\title{Calculates the exact variance}
\usage{
Fuzzy.exact.variance(data.fuzzified, breakpoints = 100, plot = FALSE)
}
\arguments{
\item{data.fuzzified}{a fuzzification matrix constructed by a call to the function FUZZ or the function GFUZZ, 
or a similar matrix. No NA are allowed.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}

\item{plot}{fixed by default to "FALSE". plot="TRUE" if a plot of the fuzzy number is required.}
}
\value{
The numerical alpha-cuts of the estimated fuzzy variance.
}
\description{
Calculates the exact variance
}
