% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_predicted_values_12122018.R
\name{fuzzy.predicted.values}
\alias{fuzzy.predicted.values}
\title{Calculates the fuzzy predicted values}
\usage{
fuzzy.predicted.values(dataset, coef.model)
}
\arguments{
\item{dataset}{the data frame containing all the variables of the model.}

\item{coef.model}{the coefficients of the model.}
}
\value{
Returns a matrix containing the alpha-cuts of the fuzzy prediced values.
}
\description{
Calculates the fuzzy predicted values
}
