% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_predicted_values_12122018.R
\name{fuzzy.residuals}
\alias{fuzzy.residuals}
\title{Calculates the fuzzy residuals}
\usage{
fuzzy.residuals(data.fuzzified, predicted.values)
}
\arguments{
\item{data.fuzzified}{the fuzzified data set constructed by a call to the function FUZZ or the function GFUZZ, or a similar matrix.}

\item{predicted.values}{the fuzzy predicted values constructed by a call to the function FUZZ or the function GFUZZ, or a similar matrix.}
}
\value{
Returns a matrix containing the alpha-cuts of the fuzzy residuals.
}
\description{
Calculates the fuzzy residuals
}
