% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy_sample_variance_17102018.R
\name{Fuzzy.sample.variance.approximation}
\alias{Fuzzy.sample.variance.approximation}
\title{Fuzzy sample variance (approx) - general}
\usage{
Fuzzy.sample.variance.approximation(data.fuzzified, appro.id)
}
\arguments{
\item{data.fuzzified}{a fuzzification matrix constructed by a call to the function FUZZ or the function GFUZZ, 
or a similar matrix. No NA are allowed.}

\item{appro.id}{an integer between 1 and 5 giving the method of approximation chosen.}
}
\value{
A numerical value.
}
\description{
Fuzzy sample variance (approx) - general
}
