% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy_sample_variance_17102018.R
\name{Fuzzy.sample.variance.approximation4}
\alias{Fuzzy.sample.variance.approximation4}
\title{Fuzzy sample variance (approx) - method 4}
\usage{
Fuzzy.sample.variance.approximation4(data.fuzzified)
}
\arguments{
\item{data.fuzzified}{a fuzzification matrix constructed by a call to the function FUZZ or the function GFUZZ, 
or a similar matrix. No NA are allowed.}
}
\value{
A numerical value.
}
\description{
Fuzzy sample variance (approx) - method 4
}
