% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuzzy_Square_26102018.R
\name{Fuzzy.Square.poly.left}
\alias{Fuzzy.Square.poly.left}
\title{Gives the polynomial expression of the left alpha-levels of the numerical square of a fuzzy number}
\usage{
Fuzzy.Square.poly.left(F1L, breakpoints = 100)
}
\arguments{
\item{F1L}{a fuzzy number.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A table containing print the related polynoms at the corresponding definition domains.
}
\description{
Gives the polynomial expression of the left alpha-levels of the numerical square of a fuzzy number
}
\examples{
X <- TrapezoidalFuzzyNumber(1,2,3,4)
Fuzzy.Square.poly.left(X)
}
