% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances_17102018.R
\name{GSGD}
\alias{GSGD}
\title{Calculates the generalized signed distance between fuzzy numbers}
\usage{
GSGD(X, Y, i = 1, j = 1, thetas = 1, breakpoints = 100)
}
\arguments{
\item{X}{a fuzzy number.}

\item{Y}{a fuzzy number.}

\item{i}{parameter of the density function of the Beta distribution, fixed by default to i = 1.}

\item{j}{parameter of the density function of the Beta distribution, fixed by default to j = 1.}

\item{thetas}{a decimal value between 0 and 1, representing the weight given to the shape of the fuzzy number. By default, thetas is fixed to 1. This parameter is used in the calculations of the d_theta star and the d_GSGD distances.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A numerical value.
}
\description{
Calculates the generalized signed distance between fuzzy numbers
}
