% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numerical_integration_18102018.R
\name{int.0}
\alias{int.0}
\title{Numerical integration by the trivial method - method 1}
\usage{
int.0(cut, a = 0, b = 1)
}
\arguments{
\item{cut}{a vector.}

\item{a}{fixed by default to 0.}

\item{b}{fixed by default to 1.}
}
\value{
An integer.
}
\description{
Numerical integration by the trivial method - method 1
}
