% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numerical_integration_18102018.R
\name{int.ct}
\alias{int.ct}
\title{Numerical integration by the composite trapezoidal method - method 3}
\usage{
int.ct(cut, a = 0, b = 1)
}
\arguments{
\item{cut}{a vector.}

\item{a}{fixed by default to 0.}

\item{b}{fixed by default to 1.}
}
\value{
An integer.
}
\description{
Numerical integration by the composite trapezoidal method - method 3
}
