% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numerical_integration_18102018.R
\name{int.simpson}
\alias{int.simpson}
\title{Numerical integration by the Simpson method - method 4}
\usage{
int.simpson(alpha, cut, a = 0, b = 1)
}
\arguments{
\item{alpha}{a vector of alpha values between 0 and 1.}

\item{cut}{a vector.}

\item{a}{fixed by default to 0.}

\item{b}{fixed by default to 1.}
}
\value{
An integer.
}
\description{
Numerical integration by the Simpson method - method 4
}
