% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numerical_integration_18102018.R
\name{integrate.num}
\alias{integrate.num}
\title{Numerical integration by a particular method}
\usage{
integrate.num(alpha, cut, method, a = 0, b = 1)
}
\arguments{
\item{alpha}{a vector of alpha values between 0 and 1.}

\item{cut}{a vector.}

\item{method}{the integration method could be one of the following four methods: "int.0", "int.t", "int.ct" and "int.simpson".}

\item{a}{fixed by default to 0.}

\item{b}{fixed by default to 1.}
}
\value{
An integer.
}
\description{
Numerical integration by a particular method
}
