% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alphacuts_17102018.R
\name{is.alphacuts}
\alias{is.alphacuts}
\title{Verifies if a matrix is set of left and right alpha-cuts}
\usage{
is.alphacuts(data)
}
\arguments{
\item{data}{a matrix of 2 equal length columns with no NA.}
}
\value{
A value TRUE if the concerned object can be a set of numerical left and right alpha-cuts, FALSE otherwise.
}
\description{
Verifies if a matrix is set of left and right alpha-cuts
}
\examples{
mat <- matrix(c(1,2,3,7,6,5), ncol = 2) 
is.alphacuts(mat)
}
