% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQ_ORDERING_13122018.R
\name{is.balanced}
\alias{is.balanced}
\title{Verifies if a design is balanced}
\usage{
is.balanced(ni)
}
\arguments{
\item{ni}{a line array given by the contingency table related to the considered variable. Often written as a result of a call of the function table.}
}
\value{
Returns a logical decision TRUE or FALSE, to indicate if a given design is respectively balanced or not.
}
\description{
Verifies if a design is balanced
}
\examples{
data <- matrix(c(1,2,3,2,2,1,1,3,1,2),ncol=1)
ni <- t(table(data))
is.balanced(ni)
}
