% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alphacuts_17102018.R
\name{is.fuzzification}
\alias{is.fuzzification}
\title{Verifies if a matrix is a fuzzification matrix}
\usage{
is.fuzzification(data)
}
\arguments{
\item{data}{an array of 3 dimensions c(m,n,2), with m lines, n columns. No NA are allowed.}
}
\value{
A value TRUE if the concerned object is a numerical fuzzification matrix, FALSE otherwise.
}
\description{
Verifies if a matrix is a fuzzification matrix
}
\examples{
mat <- array(c(1,1,2,2,3,3,5,5,6,6,7,7),dim=c(2,3,2))
is.fuzzification(mat)
}
