% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ind_Evaluation_30112018.R
\name{n_jkq.}
\alias{n_jkq.}
\title{Calculates the number of answers by a specific linguistic of a sub-item}
\usage{
n_jkq.(x, varindex, q, p_ind = rep(1, nrow(x)))
}
\arguments{
\item{x}{the data set to evaluate.}

\item{varindex}{index of a particular sub-item.}

\item{q}{index of a particular linguistic term.}

\item{p_ind}{a vector of the relative sampling weights of the units, for which \eqn{length(p_ind) = nrow(data)}. If the weights are not relative, the following expression should be applied on the vector: \deqn{\frac{p_{ind}}{\sum_{i=1}^{n} p_{ind}}.} If no sampling weights are used, the vector of weights is reduced to a vector of values 1, i.e. \eqn{rep(1, nrow(data))}.}
}
\value{
A positive integer.
}
\description{
Calculates the number of answers by a specific linguistic of a sub-item
}
