% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances_17102018.R
\name{optimal.distance}
\alias{optimal.distance}
\title{Calculates the optimal distance between two fuzzy numbers according to the chosen type}
\usage{
optimal.distance(
  X,
  Y,
  type = "DSGD.G",
  i = 1,
  j = 1,
  theta = 1/3,
  thetas = 1,
  p = 2,
  q = 0.5,
  breakpoints = 100
)
}
\arguments{
\item{X}{a fuzzy number.}

\item{Y}{a fuzzy number.}

\item{type}{type of distance chosen from the family of distances. The different choices are given by: "Rho1", "Rho2", "Bertoluzza", "Rhop", "Delta.pq", "Mid/Spr", "wabl", "DSGD", "DSGD.G", "GSGD".}

\item{i}{parameter of the density function of the Beta distribution, fixed by default to i = 1.}

\item{j}{parameter of the density function of the Beta distribution, fixed by default to j = 1.}

\item{theta}{a numerical value between 0 and 1, representing a weighting parameter. By default, theta is fixed to 1/3 referring to the Lebesgue space. This measure is used in the calculations of the following distances: d_Bertoluzza, d_mid/spr and d_phi-wabl/ldev/rdev.}

\item{thetas}{a decimal value between 0 and 1, representing the weight given to the shape of the fuzzy number. By default, thetas is fixed to 1. This parameter is used in the calculations of the d_theta star and the d_GSGD distances.}

\item{p}{a positive integer such that 1 \eqn{\le} p < infinity, referring to the parameter of the Rho_p and Delta_pq.}

\item{q}{a decimal value between 0 and 1, referring to the parameter of the metric Delta_pq.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A numerical value.
}
\description{
Calculates the optimal distance between two fuzzy numbers according to the chosen type
}
\examples{
X <- TrapezoidalFuzzyNumber(1,2,3,4) 
Y <- TrapezoidalFuzzyNumber(4,5,6,7) 
optimal.distance(X, Y, type = "GSGD")
optimal.distance(X, Y, type = "Bertoluzza")
}
