% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances_17102018.R
\name{Rho2}
\alias{Rho2}
\title{Calculates a distance by the Rho2 between fuzzy numbers}
\usage{
Rho2(X, Y, breakpoints = 100)
}
\arguments{
\item{X}{a fuzzy number.}

\item{Y}{a fuzzy number.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A numerical value.
}
\description{
Calculates a distance by the Rho2 between fuzzy numbers
}
