% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances_17102018.R
\name{Rhop}
\alias{Rhop}
\title{Calculates a distance by the d_Rhop between fuzzy numbers}
\usage{
Rhop(X, Y, p, breakpoints = 100)
}
\arguments{
\item{X}{a fuzzy number.}

\item{Y}{a fuzzy number.}

\item{p}{a positive integer such that 1 \eqn{\le} p < infinity, referring to the parameter of the Rho_p and Delta_pq.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A numerical value.
}
\description{
Calculates a distance by the d_Rhop between fuzzy numbers
}
