% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEQ_ORDERING_EXACT_19122018.R
\name{SEQ.ORDERING.EXACT}
\alias{SEQ.ORDERING.EXACT}
\title{Calculates the sequential sums of squares by an exact calculation}
\usage{
SEQ.ORDERING.EXACT(scope, data, f.response)
}
\arguments{
\item{scope}{a description of the complete fitting model.}

\item{data}{the data frame containing all the variables of the model.}

\item{f.response}{the vector of distances of the fuzzy response variable to the fuzzy origin.}
}
\value{
Returns a list of the new sets of sums of squares, as well as the coefficients, the residuals and the fitted.values.
}
\description{
Calculates the sequential sums of squares by an exact calculation
}
