% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances_17102018.R
\name{SGD}
\alias{SGD}
\title{Calculates a distance by the SGD between fuzzy numbers}
\usage{
SGD(X, i = 1, j = 1, breakpoints = 100)
}
\arguments{
\item{X}{a fuzzy number.}

\item{i}{parameter of the density function of the Beta distribution, fixed by default to i = 1.}

\item{j}{parameter of the density function of the Beta distribution, fixed by default to j = 1.}

\item{breakpoints}{a positive arbitrary integer representing the number of breaks chosen to build the numerical alpha-cuts. It is fixed to 100 by default.}
}
\value{
A numerical value.
}
\description{
Calculates a distance by the SGD between fuzzy numbers
}
